Testing the GoConvey framework
See https://github.com/smartystreets/goconvey?tab=readme-ov-file

I liked the messaging we get when we run the tests. But there is no
color when running from command line.

To make it work trough browser in a container inside the wsl proved
to be very challenging and I don't want to spend more time on it.